Add files via upload
I was appointed a data analyst for Udy_Xpress which is an e-commerce company. They have recently invested in a new technology and are keen to migrate their data into this new capability.
They currently maintain a suite of excel files, which they want to migrate onto a dedicated SQL SERVER Database. Furthermore, they want to leverage data as a source of information and will also want me  to lead on the reporting process of this project.